// src/app/question.interface.ts
export interface QuestionResult {
    response_code: number;
    results:       Question[];
}

export interface Question {
    category:          string;
    type:              string;
    difficulty:        string;
    question:          string;
    correct_answer:    string;
    incorrect_answers: string[];
}